services/runpod-worker/requirements.txt
git+https://github.com/linoytsaban/diffusers.git@wan22-loras
transformers<5
accelerate
safetensors
sentencepiece
peft
ftfy
imageio
imageio-ffmpeg
opencv-python
torchao==0.11.0
numpy>=1.16,<=1.23.5
torchvision
pillow
tqdm
huggingface_hub
runpod
pydantic==2.11.7
